@@ -4,7 +4,7 @@
   "scripts": {
     "start": "node server.js",
     "dev:start": "ng serve",
-    "dev:build": "ng build --prod",
+    "dev:build": "ng build --configuration production",
     "dev:test": "ng test",
     "dev:lint": "ng lint",
     "dev:e2e": "ng e2e",
@@ -23,19 +23,19 @@
     }
   },
   "dependencies": {
-    "@angular/animations": "11.2.5",
-    "@angular/common": "11.2.5",
-    "@angular/compiler": "11.2.5",
-    "@angular/core": "11.2.5",
-    "@angular/forms": "11.2.5",
-    "@angular/localize": "11.2.5",
-    "@angular/platform-browser": "11.2.5",
-    "@angular/platform-browser-dynamic": "11.2.5",
-    "@angular/router": "11.2.5",
+    "@angular/animations": "14.2.12",
+    "@angular/common": "14.2.12",
+    "@angular/compiler": "14.2.12",
+    "@angular/core": "14.2.12",
+    "@angular/forms": "14.2.12",
+    "@angular/localize": "14.2.12",
+    "@angular/platform-browser": "14.2.12",
+    "@angular/platform-browser-dynamic": "14.2.12",
+    "@angular/router": "14.2.12",
     "@bity/oauth2-auth-code-pkce": "^2.13.0",
-    "@ng-bootstrap/ng-bootstrap": "9.0.2",
+    "@ng-bootstrap/ng-bootstrap": "~13.0.0",
     "ansi-html": "~0.0.7",
-    "bootstrap": "4.6.0",
+    "bootstrap": "5.2.0",
     "bootstrap-select": "^1.13.18",
     "bs-stepper": "~1.7.0",
     "chalk": "~4.1.1",
@@ -50,17 +50,17 @@
     "strip-ansi": "~7.0.0",
     "tslib": "^2.0.3",
     "typescript-logger": "^5.0.1",
-    "zone.js": "0.11.4"
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.1102.14",
-    "@angular/cli": "^11.2.4",
-    "@angular/compiler-cli": "11.2.5",
-    "@angular/language-service": "11.2.5",
+    "@angular-devkit/build-angular": "^14.2.10",
+    "@angular/cli": "^14.2.10",
+    "@angular/compiler-cli": "14.2.12",
+    "@angular/language-service": "14.2.12",
     "@types/jasmine": "~3.6.2",
     "@types/jasminewd2": "2.0.8",
     "@types/node": "14.14.33",
-    "admin-lte": "3.1.0",
+    "admin-lte": "3.2.0",
     "codelyzer": "^6.0.1",
     "husky": "5.1.3",
     "jasmine-core": "~3.9.0",
@@ -78,7 +78,7 @@
     "tslint": "6.1.3",
     "tslint-config-prettier": "1.18.0",
     "tslint-plugin-prettier": "2.3.0",
-    "typescript": "4.0.6",
+    "typescript": "4.6.4",
     "webdriver-manager": "^12.1.8"
   }
 }
