@@ -1,5 +1,5 @@
 import { Component, EventEmitter, OnInit, Output, ViewChild } from '@angular/core';
-import { FormControl, FormGroup } from '@angular/forms';
+import { UntypedFormControl, UntypedFormGroup } from '@angular/forms';
 
 import { AuthService } from 'src/app/utils/services/auth.service';
 import { UserDropdownMenuComponent } from './user-dropdown-menu/user-dropdown-menu.component';
@@ -14,13 +14,13 @@ export class HeaderComponent implements OnInit {
   @ViewChild('userDropdownMenu') dropdownMenu: UserDropdownMenuComponent;
   @Output() toggleMenuSidebar: EventEmitter<any> = new EventEmitter<any>();
 
-  public searchForm: FormGroup;
+  public searchForm: UntypedFormGroup;
 
   constructor(private appService: AuthService) { }
 
   ngOnInit() {
-    this.searchForm = new FormGroup({
-      search: new FormControl(null),
+    this.searchForm = new UntypedFormGroup({
+      search: new UntypedFormControl(null),
     });
   }
 
